LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoDecoder-codec-specific-orientation.https.any_h264_annexb-expected.txt


FAIL Test that isConfigSupported() with orientation assert_equals: rotation expected (number) 90 but got (undefined) undefined
FAIL Decode frames with orientation metadata assert_equals: rotation expected (number) 90 but got (undefined) undefined
